mysql-docker/etl_mongo/requirements.txt
mysql-connector-python
pymongo
pandas
python-dateutil
